refactor sql module
